requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
together>=0.2.0
pandas>=1.5.0
openpyxl>=3.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
numpy>=1.21.0 